Game controller

+++ server/web/src/main/java/com/stockholdergame/server/web/dto/game/BuySellOperation.java
@@ -4,5 +4,5 @@ public class BuySellOperation {
 
     public Long shareId;
 
-    public Long number;
+    public int number;
 }
